Move(#12): 모임 카드 스켈레톤 폴더 변경
