+++ src/components/common/skeleton/CardSkeleton.tsx
